Update World.java
If the first operand is false (meaning the world is full) the bitwise AND operation will still check the second operand, thus adding them to the online players list. the player will never be unregistered (removed from the player list) because they they never registered. the server will always report them as online and logged in, even though this is not the case.