Add tests around CollisionManager functionality
Adds tests for the CollisionManager which makes sure that collision
matrices are correctly built for the following object types:

* Interactables
* Walls
* Corners

Missing from these tests are tests for blocked / bridged tiles, and
large corner walls, which can be added later.  Additionally, the test
cases include ASCII illustrations of the collision grid and how it is
built / checked for a particular object or movement.